[ACTP] refactor workflow runner
